extract JCloudsPropertiesBuilder into own class
